[formulas] [strategies] Move point-segment distance formula to strategies
